update from sqlite to postgres

+++ main.py
@@ -20,7 +20,6 @@ class Todo(BaseModel):
     description: Optional[str]
     priority: int = Field(gt=0, ls=6, description="Priority must between 1-5.")
     complete: bool
-    owner_id: int
 
 
 def get_db():
@@ -51,26 +50,38 @@ async def read_all(db: Session = Depends(get_db)):
 async def read_user_todos(user: dict = Depends(get_current_user),
                           db: Session = Depends(get_db)):
     if user is None:
-        raise get_user_exception
+        raise get_user_exception()
     return db.query(models.Todos).filter(models.Todos.owner_id == user.get("id")).all()
 
 
 @app.get("/todo/{todo_id}")
-async def read_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def read_todo(todo_id: int,
+                    user: dict = Depends(get_current_user),
+                    db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+
+    todo_model = db.query(models.Todos) \
+        .filter(models.Todos.id == todo_id) \
+        .filter(models.Todos.owner_id == user.get("id")) \
+        .first()
     if todo_model is not None:
         return todo_model
     raise http_notfound_exception()
 
 
 @app.post("/")
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo,
+                      user: dict = Depends(get_current_user),
+                      db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
     todo_model = models.Todos()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
-    todo_model.owner_id = todo.owner_id
+    todo_model.owner_id = user.get("id")
 
     db.add(todo_model)
     db.commit()
@@ -79,9 +90,14 @@ async def create_todo(todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.put('/{todo_id}')
-async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def update_todo(todo_id: int,
+                      todo: Todo,
+                      user: dict = Depends(get_current_user),
+                      db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
 
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
     if todo_model is None:
         raise http_notfound_exception()
 
@@ -89,6 +105,7 @@ async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get("id")
 
     db.add(todo_model)
     db.commit()
@@ -97,14 +114,20 @@ async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.delete("/{todo_id}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def delete_todo(todo_id: int,
+                      user: dict = Depends(get_current_user),
+                      db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
 
+    todo_model = db.query(models.Todos) \
+        .filter(models.Todos.id == todo_id) \
+        .filter(models.Todos.owner_id == user.get("id")) \
+        .first()
     if todo_model is None:
         raise http_notfound_exception()
 
     db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
-
     db.commit()
 
     return http_respond(HTTPStatus.OK)
